片足時のボール距離測定をマシにした. フィルタを追加. ボールの距離と方向にフィルタをかませた. NaimonにCOP(床反力中心)表示モード追加.

+++ src/jp/ac/fit/asura/nao/Sensor.java
@@ -22,7 +22,7 @@ public interface Sensor {
 
 	public float getAccelZ();
 
-	public float getForce(TouchSensor ts);
+	public int getForce(TouchSensor ts);
 
 	public float getGpsX();
